+++ dart/geobase/example/geobase_example.dart
@@ -33,8 +33,8 @@ void main() {
   _sphericalGeodesyRhumbLine();
 
   // geometric
-  _geometricCartesianPolylabel();
-  _geometricCartesianPolylabelFromGeometry();
+  _geometricCartesianPolygon();
+  _geometricCartesianPolygonFromGeometry();
 
   // geometries
   _geometryTypes2D();
@@ -650,7 +650,7 @@ void _sphericalGeodesyRhumbLine() {
   print(midPoint.latLonDms(format: dm));
 }
 
-void _geometricCartesianPolylabel() {
+void _geometricCartesianPolygon() {
   // A polygon (with an exterior ring and one interior ring as a hole) as an
   // `Iterable<PositionSeries>` that is each ring is represented by an instance
   // of `PositionSeries`.
@@ -659,17 +659,25 @@ void _geometricCartesianPolylabel() {
     [20.0, 30.0, 35.0, 35.0, 30.0, 20.0, 20.0, 30.0].positions(),
   ];
 
+  // Prints: "Centroid pos: 27.407,28.765"
+  final centroid = polygon.centroid2D();
+  print('Centroid pos: ${centroid?.toText(decimals: 3)}');
+
   // Prints: "Polylabel pos: 17.3828125,23.9453125 dist: 6.131941618102092"
   final p = polygon.polylabel2D(precision: 0.5);
   print('Polylabel pos: ${p.position} dist: ${p.distance}');
 }
 
-void _geometricCartesianPolylabelFromGeometry() {
+void _geometricCartesianPolygonFromGeometry() {
   // A polygon geometry (with an exterior ring and one interior ring as a hole).
   final polygon = Polygon.build([
     [35.0, 10.0, 45.0, 45.0, 15.0, 40.0, 10.0, 20.0, 35.0, 10.0],
     [20.0, 30.0, 35.0, 35.0, 30.0, 20.0, 20.0, 30.0],
   ]);
+
+  // Prints: "Centroid pos: 27.407,28.765"
+  final centroid = polygon.centroid2D();
+  print('Centroid pos: ${centroid?.toText(decimals: 3)}');
 
   // Prints: "Polylabel pos: 17.65625,24.21875 dist: 5.745242597140699"
   final p = polygon.polylabel2D(precision: 2.0);